Added node_migrate API
The 'node_migrate' API is used to change the cluster_id of a node.
The 'from_cluster' can be None, which means the node was an orphan node
before joining a new cluster ('to_cluster').
The 'to_cluster' can be None, which means the node is leaving its
current cluster ('from_cluster').
When both 'from_cluster' and 'to_cluster' are not None, a node is
migrated from the former one to the later one.