Make ActionProc a standalone procedure again
The previous revision to make it a class was a failure. The __call__()
function is always called without the __init__() function invoked. This
is wrong.  Also, we don't need a class wrapper for this function.

This patch also unifies context variable naming to 'context' as previous
patch suggested.  Don't use 'ctx', 'ctxt', 'cntx', 'cnxt' when 'context'
is not causing problems.