Replace 'scope' field with 'semaphore'
A second thought on this -- a single string won't be sufficient to
control access to the lock. Suppose two node actions are performed on
the same cluster, the 'NODE' scope value won't be able to tell which
action has released its lock.  This field has to be a counting
semaphore that can be the following values:
0: lock free
-1: cluster-wide action is undergoing
>0: number of node-level actions working on the cluster.