Remove cluster_lock_steal() API
This API was borrowed from Heat. I believe it is a flawed design.
Stealing lock should never be needed because it will create chaos in the
status space.

By removing this API, we need to add logics to handle engine restart
events. That is, how to handle those actions/clusters/nodes that are
locked by someone that already gone.