Remove attach/detach/update from ClusterPolicyController
In our current design, attach and detach are equivalent to the create
and delete operations on a ClusterPolicy resource.  However, it turned
out to be very cumbersome when we don't want to expose the binding IDs
for users to operate on. So we remove these operations from the
controller ClusterPolicyController and add them back to
ClusterController as action operations.