Intial version of cluster_update API support
In this implementation, we don't allow the update to the size property
because it may cause confusion in the background engine. Instead, size
should be updated using the scale-out or scale-in operations.

We also allow updates to tags, timeout, name, and parent fields of a
cluster, in addition to the 'profile_id' field. The update to cluster
properties may be type checked (e.g. timeout should be an integer).

If the profile_id is included in the update request, it may invoke a
series of CLUSTER(NODE)_UPDATE actions in the engine. We will do some
preliminary checkings in the engine side before initiating such a series
of actions.