Revised credential database table definition
This patch adds a project column to the credential table. With this new
column, we will be using (user, project) as the primary key for query.
The reason for this is that a trust may be only valid within a specified
project, while a user may be working on a different project for service
requesting.

Change-Id: Ie5f62c1d8225bbfd37fb6fed8baedd52a004e203