Revisions to cluster/node db model
This patch adds 'user' and 'domain' properties to node objects so that
they can be created with needed identity information. These identity
information will later be used to authenticate with keystone when the
physical resources are to be operated, i.e. created/update/delete.

This patch also makes the 'user', 'project', 'domain' fields fit the
current data models we get from keystone. Each of these entities has an
ID field of 32 bytes, instead of 36 bytes. For nodes or clusters, the
'user' and 'project' fields cannot be NULL.

Change-Id: If7fb248d6430bc0d499d45589fd99923b200d25e