Rename 'tags' to be 'metadata'
According to guidelines from API workgroup, we should not call the
'tags' field 'tags'. They are actually 'metadata', i.e. dictionaries
that can contain key-value pairs. 'Tags', on the other hand, is used to
refer to a list of strings. This patch fixes the DB, service, test cases
and docs to make Senlin conformant to this guideline.

Change-Id: I9692947cd02cce7f575c0bed56aa2c37e23170e4