Test cases for engine scheduler module
This patch adds unit test cases for the engine scheduler module. Maybe
we should call the 'scheduler' a 'thread_group_manager' or something
alike because it is not doing any scheduling at the moment.

This patch also deleted the 'action_wait' method which 1) is not used
anywhere 2) the implementation is buggy.

Change-Id: Ia43d26514efbfc69d935e1b26476d64f0e74d4c2