Forbid deleting profiles in use
This patch adds detection that a profile is referenced by some clusters
or some nodes. We don't allow deleting such profiles. User can do a
cluster-list or node-list with filters to find out which cluster/node is
causing problem, so we don't response with the referencing entity here.

Another tweaking is to the middleware fault module. We list exception
types in name order.

Change-Id: I72c73bae22d6ec7b819c3af5657fe3635aeecf41