Catch unexpected exception happened during action execution
Sometimes, unexpected exceptions could happened during action
execution. If these exceptions are not caught and handled
correctly, the normal progress of action execution will be
broken and the execution result and reason will not be returned
to ActionProc. In this case, other action who depends on this
action will never get the chance to finish its wait_for_dependent
progress and will hang there forever. This patch added exception
catching and handling logic in action based to resolve this problem.

Change-Id: I709c2e3283a699c67b75dfd8622fef09cbd49751
Closes-Bug: #1469956