Fix errors in trust checking
This patch revised the implementation of user id retrieval. Keystone
doesn't allow a non-admin user to perform 'users:list' operation. To
check the user id, we now check the access info from authentication
outputs.

Another problem fixed is related to trust retrieval. The 'trusts:list'
implementation in Keystone is bad. When 'trustor_user_id' is provided,
Keystone checks if the trustor_user_id is the caller; when
'trustee_user_id' is provided, Keystone checks if the trustee_user_id is
the caller. When both arguments are provided, Keystone checks if both
the trustor and the trustee is the caller, this makes no sense.

Change-Id: I1afa32aca20fd1333daf90152568c589073fc711
Closes-Bug: 1475480