Remove PolicyAction class
The PolicyAction was originally designed to be a convenient way for
enable, disable and update a policy that has been attached to a cluster.
The current implementation is flawed. We need to provide at least the
cluster ID for this operation. That makes the operation completely
redundant to the cluster_policy_update operation or
CLUSTER_POLICY_UPDATE action. This patch removes it.

Change-Id: Iec5ca536e00355733f15d11dca7df85278656aa8