Remove a TODO item in cluster detach policy
There was a TODO for checking whether there are other actions going on
when detaching a policy from a cluster. A closer examination proved that
this is unlikely to happen because CLUSTER_DETACH_POLICY action itself
is an action that will lock the cluster before getting executed. The
cluster lock already prevents other action from execution in parallel to
the curent one.

Change-Id: Ib786d688a4fe0062e06498fbfaf0c0b37c788c1e