Optimize cluster action construction
This patch moves the loading logic into the cluster action constructor.
A closer examination revealed that there is no risk that the target
cluster cannot be found (including direct action construction and those
triggered from webhook). Even there are such cases, we can leave it as a
future todo.

The intent of this change is to make cluster an attribute of the action
so that we don't have to pass it around.

Change-Id: Ic3cffa7a7f6a3b541afb2c5700aadf270bfa5c10