Add batch constraint for cluster operations
This patch adds a configuration option for creating/updating/deleting
cluster nodes. When creating a cluster, we may need to control the
amount of NODE_CREATE actions generated per-batch. However, we don't
have an opportunity to attach a policy to the cluster to control the
batch behavior. In this case, we will need a builtin batch control
mechanism so that we won't form a DOS attach on other services.

We can still attach a batch policy later to a cluster. In that case, the
batch size can be further constrained.

Change-Id: Ieaef07176bc4e94507d82cbd7d144d190c1fde30