Add DB API to get next index from cluster
This patch adds a dedicated API for getting and updating the
'next_index' that is stored as a property of a cluster. The existing
logic to update 'next_index' on node creation is problematic. We need
such an API to make sure the expected 'index' matches the index value
assigned to a node.

Change-Id: I7301ae11cc016b2ce5223c3053ff46289e4fa845