Revise node index assignment
This patch reworks the node index assignment logic which previously was
hard-coded upon node creation at db layer. In reality, we cannot
guarantee a NODE_CREATE action created ealier will get the index value
given that actions are executed in concurrently. This patch proposes a
work around so that next_index is preserved when creating a new node.

Change-Id: I0d2f4f26a681886079de82c18077d2db1093602f