Revise nova server get_details call
This patch revises the get_details call for the nova server profile. The
goal is to make the output looks more closer to that of nova show
command. Another reason is that we need to expose all attributes of a
nova server so that some policies (such as placement policy) can check
the required attributes (e.g. availability zone of a node).

Change-Id: I438a01c8910f3bfc071ff1e71389b454d0c94662