Add option to generate sample conf file.
The README says you can generate it :
https://github.com/openstack/senlin/blob/master/etc/senlin/README-senlin.conf.txt
but this option was missing.
This patch allows user to use `tox -egenconfig` to directly
generate a sample conf file in etc/senlin/senlin.conf.sample.
This helps in vetting any doubts if the already given conf is
outdated or has incorrect config options.

Change-Id: I6f8aa4cd9fd7f5296997469f04a333d67e9ddbd5