Disallow profile update for spec
Previously we support creating new profiles when doing profile-update,
if a new 'spec' is provided. However, this looks like an abuse of the
'update' operation. It creates more confusion than convenience. This
patch proposes removing 'spec' from the update interface. More patches
will come to remove this completely. This is only about the API and the
RPC client layer.

Change-Id: Id8edf1d1bcb33f85dbdad614dd4db27285f17d0a