Fix CONF.set_override usage
As mentioned in the following mail thread [1], when using CONF.set_override,
we'd better explicitly specify 'enforce_type=True', or else there is a
risk that what we are testing is not what users will be using. While the
oslo.config project may fix this param to be default to True, each
project using this call is encouraged to set it.

[1]
http://lists.openstack.org/pipermail/openstack-dev/2015-November/079633.html

Change-Id: I09f9e667cc01b6cb4b7fb5003527608a44f9a6bc