Set default auth_plugin to token when creating SDK connection
This patch revises default auth_plugin type to token when creating
SDK connection. This is because Senlin service uses token and trust
to talk with other backend openstack services. For authenticate
method in sdk driver, auth_plugin type will be decided based on
credential given. If token is not provided, auth_plugin type will be
set to password.

Change-Id: Ief88395cf7f663a501e762028566d24ffde8dfcd
Closes-Bug: #1520440