Fix status code returned from API
This patch adds a status code 'success' that represents a success
reponse a user can expect from the senlin-api. The default status code
(as implicitly specifed) in webob.Response is 200. This new status code
overrides that for a better conformance with the guidelines from the
api-wg project. When there is no guidelines for reference, we refer to
the HTTP protocol specification for definition.

Change-Id: I787739ab071bfbe70abeca8f8dc98f2e3ece8d05
partial-blueprint: http-response-modification