Remove timestamp fields from filtering parameters
The filtering logic (and grammar) today does not support complex ways
for specifying query conditions. This is something we need to improve.
However, current filter parameters still contain some timestamp fields,
on which the only filtering operation is '='. It is making these filters
not usable or even useless. This patch propose removing them. The
rationale is that once we have them in, we can NEVER deprecate them.

Change-Id: Id8ba28ef870e73994b9c20a8c2893e3e0bf18829
Closes-Bug: #1522322