Remove 'tenant_id' from senlin API paths
This is an effort to remove the 'tenant_id' from API endpoints. With
most senlin objects now augmented with 'project_safe' queries, we don't
need this 'tenant_id' in the endpoint URIs now. It is suggested that all
OS projects do this change in future.

Project based checking at API layer is not well justified a reason to
have 'tenant_id'. In fact, the check is meaningless.

Change-Id: Ia99e56e7a2a45fcdd884c502e2c03b920fee6e1a