Fix action status setting logic
This patch fixes the existing logic related to setting action status.
The 'set_status' function of the Action class is supposed to be used
when an action has finished execution. It will determin action status
based on the execution result. Previously, the function was mistakenly
used in many places when we want to signal that an action can be
started. This is terribly and sadly wrong. It is creating a lot of
problems related to action management.

This patch replaces the incorrect calls with 'action_update' DB API.

Change-Id: I09ef59686ca849da8a16bf8b9390816870befbda