Fix "report a bug" launchpad project
For docs set the launchpad project users report bugs
against to "senlin". Users can report bugs
using the "bug icon" that will
directly link to the launchpad project, it currently goes to
"openstack-manuals" which is wrong for this content.

This variable is used by openstackdocstheme 1.2.6.

Also, update comments for the variables passed to
openstackdocstheme.

Change-Id: I15d8cbc4e59b28351f21026f60b1131083540112
Related-Bug: #1524476