Fix DB coherency problem (again)
We have spent a lot time on getting the DB concurrency done right. So
far we are still getting reports that some actions' status are not
refreshed even after the database records are correctly updated. This is
very annoying.

This patch proposes a change to enforce 'expire_on_commit' so that
(hopefully) in-memory data will be expired to relfect the DB changes.

Change-Id: I9822b56ba0620125dab580d7db1773f4289fd8ec