Fix cluster resize service call
This patch fixes the problem aroused when the request is only about
chaning the min_size and/or max_size rather than a real resize
operation. This is a fix for making functional tests happy. Will add
patches for cluster_resize call in another patch. (Cannot believe we
don't have unit tests for this call).

Change-Id: I628e5865d642c3d7fd8885bc4f18b0e607f1cea6