Relax checking on server metadata
In a previous patch, we are determining the success of a join_cluster
and/or a leave_cluster operation based on whether the metadata
associated with a nova server is updated. There are some strict
checkings which are not so necessary. We should not make the senlin
engine workflow blocked by these errors. This patch proposes to relax
that checking.

Change-Id: I94d2b18a60312be0670d31d96580c82933c001a4