Reworked unit test for action service apis
We have been creating things for real in unit tests. This is a bad
behavior because it induces a huge overhead into unit tests. Another
aspect is sometimes we are testing too much. It is making the unit test
a partial functional test. This patch reworks the unit test for action
related APIs.

Change-Id: Ia8170604060b33c440b9901c9d4b0f8798ec9d44