Fix target capacity calculation
When the 'best_effort' property is set to True, we should not blindly
substract the current size by the min_size and use the result as target
capacity. There could be cases where the 'count' is less than the
difference. We should still respect the 'count' in those cases.

This patch fixes the problem.

Change-Id: I371fa352969e89a21ddc2a686c1da730e867e410