Fix action status update
When a depended action (e.g. a Child action) fails or gets cancelled, we
are only changing that action's status and the status of all dependent
actions (say, the Parent action). We also need to remove the dependency
after the action's status update. Or else, in the new design,
'action_check_status' will never report that all depended actions are
completed (failed or not).

This patch adds logic to remove the dependency item so that the
'waiting' action will eventually realize that all depended actions are
gone.

Change-Id: I735ce9f3e7b638a762341574bb44183c0b0ed695