Optimize event deserialization
When listing events or showing a particular event, we were
deserializing db records into python objects. We were doing so only to
make sure we can do a to_dict() conversion when dumping the event
properties. This is a huge waste. The oslo_db ModelBase class already
provided an _as_dict() method. We can leverage that to avoid such a
performance cost.

This patch proposes to 1) remove to_dict() method from Event class
which is a waste of time; 2) remove from_dict() method from Event which
is never used anywhere; 3) add an as_dict() method to the Event DB model
so that we still have a chance to convert datatime fields; 4) simplify
engine service layer so we can avoid doing deserialization of Events. 5)
remove all the deserialization methods from Event class.

Change-Id: I3b75632192014a60bb65b9556f9044a179d8d5c0