Set API version request on request objects
This patch adds header checking for openstack-api-version on REST API
requests. If the header is found and correctly parsed, the version
information is attached to the request before arriving at the API. If
the header doesn't exist, then a default version is attached. If the
header is specified, but is badly formatted, a 400 is returned.

Change-Id: I9db6812f1e2b86be4babb14239d33b01428782c5