Fix event generation
This patch is about a serious bug in event generation. When passing in
an action object, we have been mistakenly using the action.id as the
obj_id. This is leading to the event generated cannot be filtered using
obj_id.

We don't have unit tests for the event module yet. That will be left in
a separate patch.

Change-Id: I542b55871845956c489b1832802bdc80deac210b
Closes-Bug: 1564936