Improve parameter parsing in api-ref
This patch improves the rest_parameters parsing so that the 'optional'
property can be specified from within the RST file instead of uniformly
from the parameters.yaml file. The reason is that most parameters can be
optional in one place but be mandatory in other places. We should
specify whether a parameter is optional from where it is referenced,
rather than where it is defined (i.e. the yaml file).

Change-Id: I7449362afef8fb276c732d27a03e4dfbb650cd39