Remove emacs magic line
Some of the API doc file contain a magic line so that emacs and python
knows that the file type is ReST. Let's see if removing these lines will
create some form of panics. If we have to add them back, I'd suggest we
add them to all *.inc files, not just part of them.

Change-Id: Ibe91516116ab2b482e5e32908d39113488bac966