Initialization of versioned objects
This patch starts the work on versioned objects. The versioned objects
will server two use cases: 1) an indirection layer for the database that
lays the foundation for live upgrade in future; 2) an abstraction of
objects that can be pumped to message queue (or other channels) or
persisted into database, aka. the notification interface.

Change-Id: I2040b1cdfbe4821cdab517ed14f116e1b0f8fdc5