Fix senlin-dashboard install
When setting up tempest gate for senlin api test, we are currently
running git-clone on senlin-dashboard. It is unnecessary and it is
rejected by devstack. This patch propose a separate path for develop
setup of senlin-dashboard. Hopefully, it will not impact any production
setup while it leaves us some room for getting tempest gate up asap.

Change-Id: I314cf2e01e889a8c19fa819090ceacb98ee885eb