Rename event table columns
When adopting oslo.versionedobject, we found that obj_name has been used
as a function for VersionedObject class. Unfortunately, in senlin
database, we have 'obj_name' column (property) for events. This is
making the event objects very difficult to be serialized/deserialized.

This patch renames the 'obj_id', 'obj_name', 'obj_type' fields to 'oid',
'oname', and 'otype' respectively.

Change-Id: Idcdd979946b0e5289c4b42d2a322324d011954e0