Remove 'readonly' from property schema definition
The original intent of the 'readonly' property was to indicate whether a
property can be updated. Unfortunately, we introduced a duplicate
'updatable' propertly later which basically has the same usage. This
patch removes the 'readonly' property.

There is no user perceivable changes from this patch except for the
profile-type-show or policy-type-show. However, those two commands are
only for human readers so changing the response won't impact any
existing code. So we are not bumping API version this time.

Change-Id: I54d9674d68df5eff81872bf7f7d7d9ab0c08ff26