Replace 'format_time' with 'isotime'
After switching to interacting with DB via versioned objects, we have
datetime fields stored in ISO8601 format string. These strings, when
deserialized, contain an annoying timezone suffix which is not needed
most of the time because we are using UTC. This patch proposes to
replace format_time() calls with isotime() calls and thus remove the
former completely. For UTC date time values, the uniform representation
would be 'yyyy-mm-ddThh:MM:ssZ'.

Change-Id: I38b4d819640dff9426aa7771086a5606a0fd6219