Fix node action execution logic
We are setting cluster_id property of an orphan node to '' instead of
None, but in the node action execution logic, we are still checking if
the cluster_id is None. This patch fixes the behavior.

Another bug is related to policy checking after action execution. This
is only necessary when the action was originated from a RPC request. For
derived actions, the policy checking logic is executed from the
originating action instead of the derived one.

Change-Id: I17d84fcdaec7b26f3859c8e8118b622ebcc27189