Add 'NODE_DELETE' support to deletion policy
This patch adds the support to action 'NODE_DELETE' for deletion policy.
The patch also makes sure the 'reduce_desired_capacity' option in the
deletion policy is respected by the CLUSTER_DEL_NODES action and the
NODE_DELETE action. For the latter one, we will change desired_capacity
only if the following conditions are all met:
 1. node has a cluster ID;
 2. the action is originated from RPC request instead from a CLUSTER_xxx
    action;
 3. the 'reduce_desired_capacity' is not set to False by the policy.

Change-Id: I296dfe0c9c199a52907c7ac43b94129c010ab885