Make NODE_DELETE operation respect grace_period
This patch fixes the logic of NODE_DELETE operation when it is
originated from a RPC request. If a node is a member of a cluster, we
should respect the grace_period setting of the deletion policy, when
there exists one.

This patch also ensures that policy checking is only done when a node
action is originated from RPC. For derived actions, we don't do policy
checking twice.

Change-Id: I287512801e1f2b5d8bccd6f26b5bd24cb3c589d6