Doc for affinity and affinity policy
This patch adds user documentation about node affinity handling as
scenarios. It turns out that our affinity policy has very limited use
cases covered and there are many holes to fill considering that we allow
server group to be specified in profile specs. But anyway, we are
documenting things in the right way, and we will fix the code later.

Change-Id: I31aaf4cc71254dfb03475081f0fdd2f8f692afa4