command in doc is incorrect
When following the doc instruction I found some command is incorrect
Like: --desired-capacity <desired-capacity>
Since newer version of python-client will use --os-profile instead of
--profile, now the command with --profile is in confliction. Maybe we
should use 'senlin' for now. And change it when python-client is
ready in April 2017. In this way, the new user or contributer in
senlin will have a clear tutorial.

Change-Id: If442f78f60d6cbbc5e8f35e8e617f917a7a16efe