Fix service filtering when claiming clusters
The health manager is not claiming clusters accurately when service is
restarted. The reason is that there are some dead service records to be
removed next time one of the engines gets restarted. This patch adds a
check about the service's status and filters the service records based
on the check result.

Change-Id: Id482e74068cc1cb4394c79f14f1c6442cf9f717b
Partial-Bug: #1613552