Faithful dump nova server addresses property
Previously we were selectively dumping nova server's addresses property
in hope the output is similar to that of a 'nova show' command.
Considering that there are other properties which will be required from
the client's perspective, such as IP address version, IP address type
(i.e. 'fixed' vs. 'floating'). We are now faithfully dumping the
addresses content so that users have a chance to make a sane call.

Change-Id: I7f65f4399a9990a6ec781d9dfa0f736623f6ff77