Fix docs about autoscaling
In recent changes, we are changing the name of the metadata injected
into nova server nodes from 'cluster' to 'cluster_id'. This patch fixes
the documentation so that node grouping/aggregation can be properly
handled by the monitoring service (ceilometer/aodh).

This patch also fixes the Heat based autoscaling example where we failed
to include such a query to do proper aggregation.

Change-Id: I58762bdc3bf3fec524e4d444ed3d624e9bcba641