Fix exception handling in node_recover
The current node_recover logic (thus the profile do_recover) logic is
not raising/handling exceptions in a proper way. The type of exceptions
raised should be concise and the handling should be concise too.

This patch removes the EResourceStatusError which was never raised
anywhere in the code. It adds an EResourceOperation exception type to be
raised (in a following patch) from the driver layer. The patch also
removes the useless _handle_exception() routine from the node module.

Change-Id: I7fb2f92bf5bf5db2b5e2219704c11aa8a0f68fc7