Fix cluster status update after CLUSTER_UPDATE
Due to some historical reasons, we have been setting cluster status
based on the last operations performed on it. However, this is
incorrect. For example, a cluster_update operation may and may not
succeed, and even if it fails, it should not change the cluster's
status. The cluster's status should be always determined by its member
status.

This patch is fixing only the cluster-update operation. Other operations
should be fixed as well.

Change-Id: I2d0bf52451912c9fa98ff9846b76e9885604fff1