Remove unuse keystone interface functions
We have created quite some interface functions to shield senlin from
unexpected changes from SDK side. However, this is not so effective a
strategy so far. The recent changes 0.9.0->0.9.5 of SDK has brought many
subtle problems to both senlin and senlinclient. This patch propose
removing the interface functions we are not using at senlin side so that
we can focus on testing the ones left once SDK is making some changes.

Change-Id: I3f9ef1fe9e2a62465eb782ac0fe8a32f6978884b