Fix handling of ResourceNotFound that is not thrown
In the do_del_nodes() method, the try...except block is unnecessary
because the node_object.Node.get(...) will return 'None' if node is
not found. It will not throw an exception called 'ResourceNotFound'
in this case. This patch reworks the logic to test if the returned node
object is None.

Change-Id: I05970a995cc533f75460fd14b7e3e3ebc5be13af