Support batch policy
A batch policy allow user to perform
an action with a batch of nodes each time.
For example: a cluster contains 100 nodes.
User can assign 80 as 'min_in_service',
which means there are 80 nodes in service
no matter what action you are performing,
and then process the rest of nodes in batch.

Change-Id: Ib62639e2233c8b27a08603448d8d6f2c8e9772c6
Implements: blueprint support-batching-policy