Add http_proxy_to_wsgi to api-paste
This sets up the HTTPProxyToWSGI middleware in front of senlin-api. The
purpose of this middleware is to set up the request URL correctly in
the case there is a proxy (For instance, a loadbalancer such as HAProxy)
in front of senlin-api.

So, when TLS connections are terminated at the proxy, and one tries to
get the versions from the '/' resource from senlin-api, one will notice
that the protocol is incorrect; It will show 'http' instead of 'https'.
So this middleware handles such cases, thus helping Keystone discovery
work correctly.

The HTTPProxyToWSGI is off by default and needs to be enabled via a
configuration value.

Change-Id: Ia65ecdcc07084514884258661f840e6edcb200a4
Closes-Bug: #1590608
Co-Authored-By: pallavi <pallavi.s@nectechnologies.in>