Prepare engine and client for object parameters
We are moving away from passing JSON objects to passing versioned
objects at the RPC interface. This means the RPC client and RPC server
will need to understand (marshaling) this new data format. This patch
adds the basic serializers from oslo.versionedobject. Hope it will work
so we don't have to derive a subclass from it.

Change-Id: I1f2a6af38327d4e90727f22eead3689b7bbc3131