Add config option 'rpc_use_object'
This patch adds a new config option for switching between two versions
of RPC implementations. The option will be removed when the migration to
versioned object based parameter passing is done. The help message
indicated that so we don't want users to use it in their own setup.

This patch also renames the message topic used for health manager and
dispatcher. These two RPC servers still use the JSON serializer because
they are both senlin engine internal implementation details.

Change-Id: I6685618b7e61526bff7e2949e9a9923648c12217