Reworked cluster-create API
This reworks the cluster-create API to use the object based parameter
passing. To make things work, we had to add a 'OBJ_SERIAL_NAMESPACE'
to SenlinObject and we had to assign value other than 'versioned_object'
to it. This looks like a bug in oslo.versionedobjects. Let's at the
moment live with it and investigate it further later.

Change-Id: Id0d1b8fe3c8808b7dad245c6f0726b30a3d9a42e