Add IdentityList field type
The added 'IdentityList' field type will be used for parsing and
validating a list of entities. It is an enhanced verion of the 'List'
type from oslo.versionedobjects in that it validates whether a list has
the required minimum items, and whether the list contains duplicated
identities. The limitation is that we only accept strings for this field
type, but that should be okay for our use cases.

Change-Id: I226728dc1a30de40db836c9e40088c2f69841adb