Add TODO item about referencing existing pool
There are use cases where a user has created a pool of volumes, or a
pool of ports, or a pool of floating IPs before he/she creates a VM
cluster. When creating nodes for the cluster, he/she wants to reference
a resource from the said pool. This doesn't have to go to heat stack
profile. We are adding this to the queue so that someone in future can
pick it up.

Change-Id: I79f29c48fbe9bde1fff1d08dce9045d6e59ad6dc