Remove event logging for action signal
When signaling an action for different reasons, these are mostly not
originated from users thus not suitable for event generation or
notification. In future when we add support to action-notify, we can
override the 'signal()' method in subclasses to resurrect this
functionality. This patch removes the call to EVENT module for ease of
handling notifications in a consistent manner.

partial-blueprint: generic-event
Change-Id: I4baa22227a48b8fd6d2122ebc223c0a392578982