Simplify database driver for notification/events
Since we have eliminated the cases where we are dumping a ClusterAction
or a NodeAction directly as entity, we can focus on Cluster and Node
classes now. This patch removes the related check and handling logics.
In future we can move the entity type check logic into a parent class.

Change-Id: I03d1f675b6bc646ab19c6635fa5c57d9048efb85
partial-blueprint: generic-event