Rename node/cluster dependency key name
In senlin the relationship between nodes may have dependencies.
For example some container nodes are running on vm nodes, some nodes
are used as the DB nodes. These information are stored in the
'dependents' property of nodes which has the format like
{'nodes': ['node1', 'node2']}, at the same time the container nodes
which are running on the vm cluster, also have dependencies on vm
cluster. The profile id of container nodes are stored in cluster's
'dependents' property like {'profiles': ['profile1', 'profile']}.
This patch also fixes the bug in add/remove dependency method of db
layer, and adds more test cases to test it.

Change-Id: Ie98c7b29d8a746d70e7f4db0f82d3a2f8fbd9fd5