Add a TODO item for oslo.messaging version bump
The oslo.messaging version bump from 5.2.0 to 5.14.0 is a big thing.
Although it is not breaking our code so far, we should review the
changes to the library and see if there are things we need to change in
order to leverage its new feature and/or fix improper usage.

Change-Id: Ifcc34a44266d476e9d957f53da820704b79156e5