Make rebuild an operation for nova server
In current implementation, the 'rebuild' operation is implemented in a
profile method 'do_rebuild', this is rendering the nova server profile
so special and leaving no room for other actions to be handled.

This patch changes the 'do_rebuild' into 'handle_rebuild', which is an
operation supported by nova server profile.

Change-Id: I294cc254fda9fc6a0d92cacf8f07c5389655e399