Revise UUID field checking
We were using regular expressions for checking a UUID field. The
implementation has two problems: 1. it introduced dependency on the uuid
package, which is discouraged; 2. the logic for checking value format is
limited to 'marker' field instead of all other UUID fields.

This patch reworks the checking by using the oslo.utils.uuidutils
package.

Change-Id: Ic1c188cdcb84345764208b6b7db5d489ed91db63