Driver for glance v2
This adds a driver for glance v2. We will use this driver for checking
images. Two things to note: 1. glance v1 is deprecating, so there is no
need to add support to it; 2. nova is deprecating its proxy calls to
images, networks etc. We are supposed to use the image backend directly
in future when we need to interact with image repos.

Change-Id: Id61304dafe8f7f3e1a7de01136e4b2246e029a20