Move 'to_dict' to node object
The 'to_dict()' method was implemented in the engine.node module. This
makes the to-dict conversion very complicated. We have to use versioned
object and then the Node instance to get a dict. This is against the
whole spirit of having versioned object to abstract database details.

This patch moves the method to versioned object 'node.Node'.

Change-Id: Ie6b097936e1d78c3e79268994d6caa942e0008c1