Fix node deletion logic
This is an optimization of the node deletion logic. Previously there was
a TODO item about checking if there are actions still running on the
node. A careful reasoning revealed that it is impossible. The node
delete logic has to start with a NODE_DELETE action which is already
locking the node. No other action could grab the lock.

Change-Id: I8a28164861c2a639d7de09c6a6a2c0ad6af4e878