Use dynamic timer for cluster status polling
In the current implementation, we have a fixed-interval timer for
polling each cluster when appropriate. However, this could be very
dangerous because we cannot guarantee that the polling will complete
during the given interval. This situation is worsened when we later
added the triggering of NODE_RECOVER actions.

For example, if the polling interval is specified as 2 seconds, it
means that the user wants the cluster health to be triggered every 2
seconds. Even if a previous checking is still ongoing, the next checking
loop will be fired off nonetheless.

This patch fixes this situation by using "dynamic timers". A timer is
fired with a callback function to be executed (_poll_cluster in this
case). The function is supposed to check if it has taken a longer time
to complete its job. A warning message will be emitted if so (possible
improvement is to fire an event? ) Suppose the _poll_cluster operation
has taken 5 seconds to finish. The expected behavior is now:

 - A warning message is logged saying that the poller has missed two
   intervals.

 - The poller will return a sugguested time for sleep, which is 1 in
   this case.

After sleeping for 1 second, the _poll_cluster function will be invoked
again as if nothing has happened. In the revised logic, we are no longer
guaranteeing that the poller will always be triggered. But whenever we
failed to do a poll in time, we log a warning.

Change-Id: I7d894ee55d386670281610abefbd69898a172d9b