Fixed inappropriate event generation in action base
When policy check succeeds, we are not supposed to emit an event (either
to database or to message queue) because the success of such a check is
not interesting to users at all. They only care about the abnormal
behaviors, aka. errors or warnings.

Change-Id: I4cfbb8fb90aeefc3750365c9d3998bd3246c21d1
(cherry picked from commit 340a0d3b9e6f97ca2832eb575b0a4993330a9419)