Improve the slow path of lock acquire
The slow path of cluster_lock_acquire was implemented as only marking
the related action as a failure when stealing locks from a dead engine.
This is problematic. We are not sure if the dead engine has also
impacted other objects, e.g. nodes.

This patch replaces that cleansing call with a more thorough solution.

Change-Id: I43458f2a9c6d709d670b36cbd2ab538a8dbc8a2e