Fix LB member status update in error conditions
When we fail in adding a node to the LB pool or removing a node from the
LB pool, we should set the node status to WARNING. This is a hint to
users that the node is not supposed to function as expected.

This patch also fixes the logic about updating action status when
checking the LB policy.

Change-Id: I0857c35782b42b80e8f0ef8759fb7d144a7813a8