60 seconds for a cluster check seems too short
When a health policy was created with the value in the
example, it will run check each 60 seconds, it will gererate
many checking events and actions.
And when the node number in cluster is a bit large it always
can't complete cluster check in 60 seconds.
So maybe it's better to set the default example interval to 600 seconds.

Change-Id: I272ab7e3001cbd8e3640ef15d34e12b48442065f