Enable more inclusive cluster filtering when listing events
This patch adds logic so that when listing events with cluster 'id' as
filters, users can actually specify the name, full-id or short-id of a
cluster. It will be more convenient for users.
If any of the specified cluster is not found from the database, we
return an empty list of events.

Closes-Bug: #1642869
Change-Id: Ic27053e6b90658f35b577038d26b4aebb9d4475f