Remove a TODO item related to oslo.messaging
Previously, we left a TODO item related to the version bump (a big one)
from oslo.messaging because we are not so sure there is nothing left to
be done for Senlin to adapt to such a change.
This patch is a double confirmation that the version bump is mostly
about oslo.messaing internals. We do NOT need to do any changes
accordingly.

Change-Id: I47497609ca9d8ff980f8881d4fc572664941bf87